requirements.txt
pypdf==4.2.0
chromadb==0.5.0
google-generativeai==0.6.0
python-dotenv==1.0.1
